refactored image preprocessing
