--- clip/image_constants.py
@@ -1,5 +1,0 @@
-from typing import List, Union
-
-ImageInput = Union[
-    "PIL.Image.Image", List["PIL.Image.Image"]
-]
