--- clip/image_processor.py
@@ -1,252 +1,0 @@
-from image_config import CLIP_IMAGE_CONFIG
-from jmage_constants import ImageInput
-from image_utils import get_size_dict, infer_channel_dimension_format, get_resize_output_image_size, ChannelDimension, rescale, to_channel_dimension_format
-from image_transforms import convert_to_rgb, to_numpy_array, resize, center_crop, normalize
-import numpy as np
-import PIL
-from typing import Dict, Optional, Union, Iterable
-import mlx.core as mx
-
-
-class CLIPImageProcessor:
-    def __init__(self, model_name):
-        self.model_name = model_name
-
-        self.config = CLIP_IMAGE_CONFIG[model_name]
-
-        self.do_resize = self.config["do_resize"]
-        size = self.config["size"]
-        self.size = get_size_dict(
-            size, param_name="size", default_to_square=False)
-        self.resample = self.config["resample"]
-        self.do_center_crop = self.config["do_center_crop"]
-        crop_size = self.config["crop_size"]
-        self.crop_size = get_size_dict(
-            crop_size, param_name="crop_size", default_to_square=True)
-        self.do_rescale = self.config["do_rescale"]
-        self.rescale_factor = self.config["rescale_factor"]
-        self.do_normalize = self.config["do_normalize"]
-        self.image_mean = self.config["image_mean"]
-        self.image_std = self.config["image_std"]
-        self.do_convert_rgb = self.config["do_convert_rgb"]
-
-    def __call__(self, images: ImageInput, data_format: Optional[Union[str, ChannelDimension]] = ChannelDimension.FIRST):
-        """
-        Converts a single or list of images into the proper MLX format
-        """
-
-        if not isinstance(images, list):
-            images = [images]
-
-        # PIL RGBA images are converted to RGB
-        if self.do_convert_rgb:
-            images = [convert_to_rgb(image) for image in images]
-
-        # All transformations expect numpy arrays.
-        images = [to_numpy_array(image) for image in images]
-
-        input_data_format = infer_channel_dimension_format(images[0])
-
-        if self.do_resize:
-            images = [
-                self.resize_image(image=image, size=self.size,
-                                  resample=self.resample, input_data_format=input_data_format)
-                for image in images
-            ]
-
-        if self.do_center_crop:
-            images = [
-                self.center_crop_image(image=image, size=self.crop_size, input_data_format=input_data_format) for image in images
-            ]
-
-        if self.do_rescale:
-            images = [
-                self.rescale_image(
-                    image=image, scale=self.rescale_factor, input_data_format=input_data_format)
-                for image in images
-            ]
-
-        if self.do_normalize:
-            images = [
-                self.normalize_image(image=image, mean=self.image_mean,
-                                     std=self.image_std, input_data_format=input_data_format)
-                for image in images
-            ]
-
-        images = [
-            to_channel_dimension_format(image, data_format, input_channel_dim=input_data_format) for image in images
-        ]
-
-        data = {"pixel_values": mx.array(
-            np.array(images)).transpose((0, 2, 3, 1))}
-
-        return data
-
-    def resize_image(
-        self,
-        image: np.ndarray,
-        size: Dict[str, int],
-        resample=PIL.Image.BICUBIC,
-        data_format: Optional[Union[str, ChannelDimension]] = None,
-        input_data_format: Optional[Union[str, ChannelDimension]] = None,
-        **kwargs,
-    ) -> np.ndarray:
-        """
-        Resize an image. The shortest edge of the image is resized to size["shortest_edge"], with the longest edge
-        resized to keep the input aspect ratio.
-
-        Args:
-            image (`np.ndarray`):
-                Image to resize.
-            size (`Dict[str, int]`):
-                Size of the output image.
-            resample (`PILImageResampling`, *optional*, defaults to `PILImageResampling.BICUBIC`):
-                Resampling filter to use when resiizing the image.
-            data_format (`str` or `ChannelDimension`, *optional*):
-                The channel dimension format of the image. If not provided, it will be the same as the input image.
-            input_data_format (`ChannelDimension` or `str`, *optional*):
-                The channel dimension format of the input image. If not provided, it will be inferred.
-        """
-        default_to_square = True
-        if "shortest_edge" in size:
-            size = size["shortest_edge"]
-            default_to_square = False
-        elif "height" in size and "width" in size:
-            size = (size["height"], size["width"])
-        else:
-            raise ValueError(
-                "Size must contain either 'shortest_edge' or 'height' and 'width'.")
-
-        output_size = get_resize_output_image_size(
-            image,
-            size=size,
-            default_to_square=default_to_square,
-            input_data_format=input_data_format,
-        )
-        return resize(
-            image,
-            size=output_size,
-            resample=resample,
-            data_format=data_format,
-            input_data_format=input_data_format,
-            **kwargs,
-        )
-
-    def center_crop_image(
-        self,
-        image: np.ndarray,
-        size: Dict[str, int],
-        data_format: Optional[Union[str, ChannelDimension]] = None,
-        input_data_format: Optional[Union[str, ChannelDimension]] = None,
-        **kwargs,
-    ) -> np.ndarray:
-        """
-        Center crop an image to `(size["height"], size["width"])`. If the input size is smaller than `crop_size` along
-        any edge, the image is padded with 0's and then center cropped.
-
-        Args:
-            image (`np.ndarray`):
-                Image to center crop.
-            size (`Dict[str, int]`):
-                Size of the output image.
-            data_format (`str` or `ChannelDimension`, *optional*):
-                The channel dimension format for the output image. If unset, the channel dimension format of the input
-                image is used. Can be one of:
-                - `"channels_first"` or `ChannelDimension.FIRST`: image in (num_channels, height, width) format.
-                - `"channels_last"` or `ChannelDimension.LAST`: image in (height, width, num_channels) format.
-            input_data_format (`ChannelDimension` or `str`, *optional*):
-                The channel dimension format for the input image. If unset, the channel dimension format is inferred
-                from the input image. Can be one of:
-                - `"channels_first"` or `ChannelDimension.FIRST`: image in (num_channels, height, width) format.
-                - `"channels_last"` or `ChannelDimension.LAST`: image in (height, width, num_channels) format.
-        """
-        size = get_size_dict(size)
-        if "height" not in size or "width" not in size:
-            raise ValueError(
-                f"The size dictionary must have keys 'height' and 'width'. Got {size.keys()}")
-        return center_crop(
-            image,
-            size=(size["height"], size["width"]),
-            data_format=data_format,
-            input_data_format=input_data_format,
-            **kwargs,
-        )
-
-    def rescale_image(
-        self,
-        image: np.ndarray,
-        scale: float,
-        data_format: Optional[Union[str, ChannelDimension]] = None,
-        input_data_format: Optional[Union[str, ChannelDimension]] = None,
-        **kwargs,
-    ) -> np.ndarray:
-        """
-        Rescale an image by a scale factor. image = image * scale.
-
-        Args:
-            image (`np.ndarray`):
-                Image to rescale.
-            scale (`float`):
-                The scaling factor to rescale pixel values by.
-            data_format (`str` or `ChannelDimension`, *optional*):
-                The channel dimension format for the output image. If unset, the channel dimension format of the input
-                image is used. Can be one of:
-                - `"channels_first"` or `ChannelDimension.FIRST`: image in (num_channels, height, width) format.
-                - `"channels_last"` or `ChannelDimension.LAST`: image in (height, width, num_channels) format.
-            input_data_format (`ChannelDimension` or `str`, *optional*):
-                The channel dimension format for the input image. If unset, the channel dimension format is inferred
-                from the input image. Can be one of:
-                - `"channels_first"` or `ChannelDimension.FIRST`: image in (num_channels, height, width) format.
-                - `"channels_last"` or `ChannelDimension.LAST`: image in (height, width, num_channels) format.
-
-        Returns:
-            `np.ndarray`: The rescaled image.
-        """
-        return rescale(image, scale=scale, data_format=data_format, input_data_format=input_data_format, **kwargs)
-
-    def normalize_image(
-        self,
-        image: np.ndarray,
-        mean: Union[float, Iterable[float]],
-        std: Union[float, Iterable[float]],
-        data_format: Optional[Union[str, ChannelDimension]] = None,
-        input_data_format: Optional[Union[str, ChannelDimension]] = None,
-        **kwargs,
-    ) -> np.ndarray:
-        """
-        Normalize an image. image = (image - image_mean) / image_std.
-
-        Args:
-            image (`np.ndarray`):
-                Image to normalize.
-            mean (`float` or `Iterable[float]`):
-                Image mean to use for normalization.
-            std (`float` or `Iterable[float]`):
-                Image standard deviation to use for normalization.
-            data_format (`str` or `ChannelDimension`, *optional*):
-                The channel dimension format for the output image. If unset, the channel dimension format of the input
-                image is used. Can be one of:
-                - `"channels_first"` or `ChannelDimension.FIRST`: image in (num_channels, height, width) format.
-                - `"channels_last"` or `ChannelDimension.LAST`: image in (height, width, num_channels) format.
-            input_data_format (`ChannelDimension` or `str`, *optional*):
-                The channel dimension format for the input image. If unset, the channel dimension format is inferred
-                from the input image. Can be one of:
-                - `"channels_first"` or `ChannelDimension.FIRST`: image in (num_channels, height, width) format.
-                - `"channels_last"` or `ChannelDimension.LAST`: image in (height, width, num_channels) format.
-
-        Returns:
-            `np.ndarray`: The normalized image.
-        """
-        return normalize(
-            image, mean=mean, std=std, data_format=data_format, input_data_format=input_data_format, **kwargs
-        )
-
-
-if __name__ == "__main__":
-    mx_processor = CLIPImageProcessor("openai/clip-vit-base-patch32")
-
-    image = PIL.Image.open("cats.jpeg")
-
-    data = mx_processor(image)
-
-    print(data)
