--- clip/image_transforms.py
@@ -1,269 +1,0 @@
-from typing import Iterable, Optional, Tuple, Union
-from image_utils import ChannelDimension, infer_channel_dimension_format, to_channel_dimension_format, _rescale_for_pil_conversion, to_pil_image, rescale, get_image_size, get_channel_dimension_axis
-import PIL
-from image_processor_constants import ImageInput
-import numpy as np
-from PIL import Image
-
-
-def convert_to_rgb(image: ImageInput) -> ImageInput:
-    """
-    Converts an image to RGB format. Only converts if the image is of type PIL.Image.Image, otherwise returns the image
-    as is.
-
-    Args:
-        image (Image):
-            The image to convert.
-    """
-
-    if not isinstance(image, PIL.Image.Image):
-        return image
-
-    image = image.convert("RGB")
-    return image
-
-
-def to_numpy_array(img) -> np.ndarray:
-    if isinstance(img, Image.Image):
-        return np.array(img)
-    elif isinstance(img, np.ndarray):
-        return img
-    else:
-        raise ValueError(f"Unsupported type {type(img)}")
-
-
-def resize(
-    image: np.ndarray,
-    size: Tuple[int, int],
-    resample: "PILImageResampling" = None,
-    reducing_gap: Optional[int] = None,
-    data_format: Optional[ChannelDimension] = None,
-    return_numpy: bool = True,
-    input_data_format: Optional[Union[str, ChannelDimension]] = None,
-) -> np.ndarray:
-    """
-    Resizes `image` to `(height, width)` specified by `size` using the PIL library.
-
-    Args:
-        image (`np.ndarray`):
-            The image to resize.
-        size (`Tuple[int, int]`):
-            The size to use for resizing the image.
-        resample (`int`, *optional*, defaults to `PILImageResampling.BILINEAR`):
-            The filter to user for resampling.
-        reducing_gap (`int`, *optional*):
-            Apply optimization by resizing the image in two steps. The bigger `reducing_gap`, the closer the result to
-            the fair resampling. See corresponding Pillow documentation for more details.
-        data_format (`ChannelDimension`, *optional*):
-            The channel dimension format of the output image. If unset, will use the inferred format from the input.
-        return_numpy (`bool`, *optional*, defaults to `True`):
-            Whether or not to return the resized image as a numpy array. If False a `PIL.Image.Image` object is
-            returned.
-        input_data_format (`ChannelDimension`, *optional*):
-            The channel dimension format of the input image. If unset, will use the inferred format from the input.
-
-    Returns:
-        `np.ndarray`: The resized image.
-    """
-
-    if not len(size) == 2:
-        raise ValueError("size must have 2 elements")
-
-    # For all transformations, we want to keep the same data format as the input image unless otherwise specified.
-    # The resized image from PIL will always have channels last, so find the input format first.
-    if input_data_format is None:
-        input_data_format = infer_channel_dimension_format(image)
-    data_format = input_data_format if data_format is None else data_format
-
-    # To maintain backwards compatibility with the resizing done in previous image feature extractors, we use
-    # the pillow library to resize the image and then convert back to numpy
-    do_rescale = False
-    if not isinstance(image, PIL.Image.Image):
-        do_rescale = _rescale_for_pil_conversion(image)
-        image = to_pil_image(image, do_rescale=do_rescale,
-                             input_data_format=input_data_format)
-    height, width = size
-    # PIL images are in the format (width, height)
-    resized_image = image.resize(
-        (width, height), resample=resample, reducing_gap=reducing_gap)
-
-    if return_numpy:
-        resized_image = np.array(resized_image)
-        # If the input image channel dimension was of size 1, then it is dropped when converting to a PIL image
-        # so we need to add it back if necessary.
-        resized_image = np.expand_dims(
-            resized_image, axis=-1) if resized_image.ndim == 2 else resized_image
-        # The image is always in channels last format after converting from a PIL image
-        resized_image = to_channel_dimension_format(
-            resized_image, data_format, input_channel_dim=ChannelDimension.LAST
-        )
-        # If an image was rescaled to be in the range [0, 255] before converting to a PIL image, then we need to
-        # rescale it back to the original range.
-        resized_image = rescale(
-            resized_image, 1 / 255) if do_rescale else resized_image
-    return resized_image
-
-
-def center_crop(
-    image: np.ndarray,
-    size: Tuple[int, int],
-    data_format: Optional[Union[str, ChannelDimension]] = None,
-    input_data_format: Optional[Union[str, ChannelDimension]] = None,
-    return_numpy: Optional[bool] = None,
-) -> np.ndarray:
-    """
-    Crops the `image` to the specified `size` using a center crop. Note that if the image is too small to be cropped to
-    the size given, it will be padded (so the returned result will always be of size `size`).
-
-    Args:
-        image (`np.ndarray`):
-            The image to crop.
-        size (`Tuple[int, int]`):
-            The target size for the cropped image.
-        data_format (`str` or `ChannelDimension`, *optional*):
-            The channel dimension format for the output image. Can be one of:
-                - `"channels_first"` or `ChannelDimension.FIRST`: image in (num_channels, height, width) format.
-                - `"channels_last"` or `ChannelDimension.LAST`: image in (height, width, num_channels) format.
-            If unset, will use the inferred format of the input image.
-        input_data_format (`str` or `ChannelDimension`, *optional*):
-            The channel dimension format for the input image. Can be one of:
-                - `"channels_first"` or `ChannelDimension.FIRST`: image in (num_channels, height, width) format.
-                - `"channels_last"` or `ChannelDimension.LAST`: image in (height, width, num_channels) format.
-            If unset, will use the inferred format of the input image.
-        return_numpy (`bool`, *optional*):
-            Whether or not to return the cropped image as a numpy array. Used for backwards compatibility with the
-            previous ImageFeatureExtractionMixin method.
-                - Unset: will return the same type as the input image.
-                - `True`: will return a numpy array.
-                - `False`: will return a `PIL.Image.Image` object.
-    Returns:
-        `np.ndarray`: The cropped image.
-    """
-
-    return_numpy = True if return_numpy is None else return_numpy
-
-    if not isinstance(image, np.ndarray):
-        raise ValueError(
-            f"Input image must be of type np.ndarray, got {type(image)}")
-
-    if not isinstance(size, Iterable) or len(size) != 2:
-        raise ValueError(
-            "size must have 2 elements representing the height and width of the output image")
-
-    if input_data_format is None:
-        input_data_format = infer_channel_dimension_format(image)
-    output_data_format = data_format if data_format is not None else input_data_format
-
-    # We perform the crop in (C, H, W) format and then convert to the output format
-    image = to_channel_dimension_format(
-        image, ChannelDimension.FIRST, input_data_format)
-
-    orig_height, orig_width = get_image_size(image, ChannelDimension.FIRST)
-    crop_height, crop_width = size
-    crop_height, crop_width = int(crop_height), int(crop_width)
-
-    # In case size is odd, (image_shape[0] + size[0]) // 2 won't give the proper result.
-    top = (orig_height - crop_height) // 2
-    bottom = top + crop_height
-    # In case size is odd, (image_shape[1] + size[1]) // 2 won't give the proper result.
-    left = (orig_width - crop_width) // 2
-    right = left + crop_width
-
-    # Check if cropped area is within image boundaries
-    if top >= 0 and bottom <= orig_height and left >= 0 and right <= orig_width:
-        image = image[..., top:bottom, left:right]
-        image = to_channel_dimension_format(
-            image, output_data_format, ChannelDimension.FIRST)
-        return image
-
-    # Otherwise, we may need to pad if the image is too small. Oh joy...
-    new_height = max(crop_height, orig_height)
-    new_width = max(crop_width, orig_width)
-    new_shape = image.shape[:-2] + (new_height, new_width)
-    new_image = np.zeros_like(image, shape=new_shape)
-
-    # If the image is too small, pad it with zeros
-    top_pad = (new_height - orig_height) // 2
-    bottom_pad = top_pad + orig_height
-    left_pad = (new_width - orig_width) // 2
-    right_pad = left_pad + orig_width
-    new_image[..., top_pad:bottom_pad, left_pad:right_pad] = image
-
-    top += top_pad
-    bottom += top_pad
-    left += left_pad
-    right += left_pad
-
-    new_image = new_image[..., max(0, top): min(
-        new_height, bottom), max(0, left): min(new_width, right)]
-    new_image = to_channel_dimension_format(
-        new_image, output_data_format, ChannelDimension.FIRST)
-
-    if not return_numpy:
-        new_image = to_pil_image(new_image)
-
-    return new_image
-
-
-def normalize(
-    image: np.ndarray,
-    mean: Union[float, Iterable[float]],
-    std: Union[float, Iterable[float]],
-    data_format: Optional[ChannelDimension] = None,
-    input_data_format: Optional[Union[str, ChannelDimension]] = None,
-) -> np.ndarray:
-    """
-    Normalizes `image` using the mean and standard deviation specified by `mean` and `std`.
-
-    image = (image - mean) / std
-
-    Args:
-        image (`np.ndarray`):
-            The image to normalize.
-        mean (`float` or `Iterable[float]`):
-            The mean to use for normalization.
-        std (`float` or `Iterable[float]`):
-            The standard deviation to use for normalization.
-        data_format (`ChannelDimension`, *optional*):
-            The channel dimension format of the output image. If unset, will use the inferred format from the input.
-        input_data_format (`ChannelDimension`, *optional*):
-            The channel dimension format of the input image. If unset, will use the inferred format from the input.
-    """
-    if not isinstance(image, np.ndarray):
-        raise ValueError("image must be a numpy array")
-
-    if input_data_format is None:
-        input_data_format = infer_channel_dimension_format(image)
-    channel_axis = get_channel_dimension_axis(
-        image, input_data_format=input_data_format)
-    num_channels = image.shape[channel_axis]
-
-    # We cast to float32 to avoid errors that can occur when subtracting uint8 values.
-    # We preserve the original dtype if it is a float type to prevent upcasting float16.
-    if not np.issubdtype(image.dtype, np.floating):
-        image = image.astype(np.float32)
-
-    if isinstance(mean, Iterable):
-        if len(mean) != num_channels:
-            raise ValueError(
-                f"mean must have {num_channels} elements if it is an iterable, got {len(mean)}")
-    else:
-        mean = [mean] * num_channels
-    mean = np.array(mean, dtype=image.dtype)
-
-    if isinstance(std, Iterable):
-        if len(std) != num_channels:
-            raise ValueError(
-                f"std must have {num_channels} elements if it is an iterable, got {len(std)}")
-    else:
-        std = [std] * num_channels
-    std = np.array(std, dtype=image.dtype)
-
-    if input_data_format == ChannelDimension.LAST:
-        image = (image - mean) / std
-    else:
-        image = ((image.T - mean) / std).T
-
-    image = to_channel_dimension_format(
-        image, data_format, input_data_format) if data_format is not None else image
-    return image
