--- clip/image_utils.py
@@ -1,483 +1,0 @@
-from typing import Dict, Iterable, List, Optional, Tuple, Union
-from enum import Enum
-import numpy as np
-import PIL
-
-VALID_SIZE_DICT_KEYS = ({"height", "width"}, {"shortest_edge"}, {
-                        "shortest_edge", "longest_edge"}, {"longest_edge"})
-
-
-def is_valid_size_dict(size_dict):
-    if not isinstance(size_dict, dict):
-        return False
-
-    size_dict_keys = set(size_dict.keys())
-    for allowed_keys in VALID_SIZE_DICT_KEYS:
-        if size_dict_keys == allowed_keys:
-            return True
-    return False
-
-
-def convert_to_size_dict(
-    size, max_size: Optional[int] = None, default_to_square: bool = True, height_width_order: bool = True
-):
-    # By default, if size is an int we assume it represents a tuple of (size, size).
-    if isinstance(size, int) and default_to_square:
-        if max_size is not None:
-            raise ValueError(
-                "Cannot specify both size as an int, with default_to_square=True and max_size")
-        return {"height": size, "width": size}
-    # In other configs, if size is an int and default_to_square is False, size represents the length of
-    # the shortest edge after resizing.
-    elif isinstance(size, int) and not default_to_square:
-        size_dict = {"shortest_edge": size}
-        if max_size is not None:
-            size_dict["longest_edge"] = max_size
-        return size_dict
-    # Otherwise, if size is a tuple it's either (height, width) or (width, height)
-    elif isinstance(size, (tuple, list)) and height_width_order:
-        return {"height": size[0], "width": size[1]}
-    elif isinstance(size, (tuple, list)) and not height_width_order:
-        return {"height": size[1], "width": size[0]}
-    elif size is None and max_size is not None:
-        if default_to_square:
-            raise ValueError(
-                "Cannot specify both default_to_square=True and max_size")
-        return {"longest_edge": max_size}
-
-    raise ValueError(f"Could not convert size input to size dict: {size}")
-
-
-def get_size_dict(
-    size: Union[int, Iterable[int], Dict[str, int]] = None,
-    max_size: Optional[int] = None,
-    height_width_order: bool = True,
-    default_to_square: bool = True,
-    param_name="size",
-) -> dict:
-    """
-    Converts the old size parameter in the config into the new dict expected in the config. This is to ensure backwards
-    compatibility with the old image processor configs and removes ambiguity over whether the tuple is in (height,
-    width) or (width, height) format.
-
-    - If `size` is tuple, it is converted to `{"height": size[0], "width": size[1]}` or `{"height": size[1], "width":
-    size[0]}` if `height_width_order` is `False`.
-    - If `size` is an int, and `default_to_square` is `True`, it is converted to `{"height": size, "width": size}`.
-    - If `size` is an int and `default_to_square` is False, it is converted to `{"shortest_edge": size}`. If `max_size`
-      is set, it is added to the dict as `{"longest_edge": max_size}`.
-
-    Args:
-        size (`Union[int, Iterable[int], Dict[str, int]]`, *optional*):
-            The `size` parameter to be cast into a size dictionary.
-        max_size (`Optional[int]`, *optional*):
-            The `max_size` parameter to be cast into a size dictionary.
-        height_width_order (`bool`, *optional*, defaults to `True`):
-            If `size` is a tuple, whether it's in (height, width) or (width, height) order.
-        default_to_square (`bool`, *optional*, defaults to `True`):
-            If `size` is an int, whether to default to a square image or not.
-    """
-    if not isinstance(size, dict):
-        size_dict = convert_to_size_dict(
-            size, max_size, default_to_square, height_width_order)
-        print(
-            f"{param_name} should be a dictionary on of the following set of keys: {VALID_SIZE_DICT_KEYS}, got {size}."
-            f" Converted to {size_dict}.",
-        )
-    else:
-        size_dict = size
-
-    if not is_valid_size_dict(size_dict):
-        raise ValueError(
-            f"{param_name} must have one of the following set of keys: {VALID_SIZE_DICT_KEYS}, got {size_dict.keys()}"
-        )
-    return size_dict
-
-
-def get_size_dict(
-    size: Union[int, Iterable[int], Dict[str, int]] = None,
-    max_size: Optional[int] = None,
-    height_width_order: bool = True,
-    default_to_square: bool = True,
-    param_name="size",
-) -> dict:
-    """
-    Converts the old size parameter in the config into the new dict expected in the config. This is to ensure backwards
-    compatibility with the old image processor configs and removes ambiguity over whether the tuple is in (height,
-    width) or (width, height) format.
-
-    - If `size` is tuple, it is converted to `{"height": size[0], "width": size[1]}` or `{"height": size[1], "width":
-    size[0]}` if `height_width_order` is `False`.
-    - If `size` is an int, and `default_to_square` is `True`, it is converted to `{"height": size, "width": size}`.
-    - If `size` is an int and `default_to_square` is False, it is converted to `{"shortest_edge": size}`. If `max_size`
-      is set, it is added to the dict as `{"longest_edge": max_size}`.
-
-    Args:
-        size (`Union[int, Iterable[int], Dict[str, int]]`, *optional*):
-            The `size` parameter to be cast into a size dictionary.
-        max_size (`Optional[int]`, *optional*):
-            The `max_size` parameter to be cast into a size dictionary.
-        height_width_order (`bool`, *optional*, defaults to `True`):
-            If `size` is a tuple, whether it's in (height, width) or (width, height) order.
-        default_to_square (`bool`, *optional*, defaults to `True`):
-            If `size` is an int, whether to default to a square image or not.
-    """
-    if not isinstance(size, dict):
-        size_dict = convert_to_size_dict(
-            size, max_size, default_to_square, height_width_order)
-        print(
-            f"{param_name} should be a dictionary on of the following set of keys: {VALID_SIZE_DICT_KEYS}, got {size}."
-            f" Converted to {size_dict}.",
-        )
-    else:
-        size_dict = size
-
-    if not is_valid_size_dict(size_dict):
-        raise ValueError(
-            f"{param_name} must have one of the following set of keys: {VALID_SIZE_DICT_KEYS}, got {size_dict.keys()}"
-        )
-    return size_dict
-
-
-class ExplicitEnum(str, Enum):
-    """
-    Enum with more explicit error message for missing values.
-    """
-
-    @classmethod
-    def _missing_(cls, value):
-        raise ValueError(
-            f"{value} is not a valid {cls.__name__}, please select one of {list(cls._value2member_map_.keys())}"
-        )
-
-
-class ChannelDimension(ExplicitEnum):
-    FIRST = "channels_first"
-    LAST = "channels_last"
-
-
-def infer_channel_dimension_format(
-    image: np.ndarray, num_channels: Optional[Union[int, Tuple[int, ...]]] = None
-) -> ChannelDimension:
-    """
-    Infers the channel dimension format of `image`.
-
-    Args:
-        image (`np.ndarray`):
-            The image to infer the channel dimension of.
-        num_channels (`int` or `Tuple[int, ...]`, *optional*, defaults to `(1, 3)`):
-            The number of channels of the image.
-
-    Returns:
-        The channel dimension of the image.
-    """
-    num_channels = num_channels if num_channels is not None else (1, 3)
-    num_channels = (num_channels,) if isinstance(
-        num_channels, int) else num_channels
-
-    if image.ndim == 3:
-        first_dim, last_dim = 0, 2
-    elif image.ndim == 4:
-        first_dim, last_dim = 1, 3
-    else:
-        raise ValueError(
-            f"Unsupported number of image dimensions: {image.ndim}")
-
-    if image.shape[first_dim] in num_channels:
-        return ChannelDimension.FIRST
-    elif image.shape[last_dim] in num_channels:
-        return ChannelDimension.LAST
-    raise ValueError("Unable to infer channel dimension format")
-
-
-def get_resize_output_image_size(
-    input_image: np.ndarray,
-    size: Union[int, Tuple[int, int], List[int], Tuple[int]],
-    default_to_square: bool = True,
-    max_size: Optional[int] = None,
-    input_data_format: Optional[Union[str, ChannelDimension]] = None,
-) -> tuple:
-    """
-    Find the target (height, width) dimension of the output image after resizing given the input image and the desired
-    size.
-
-    Args:
-        input_image (`np.ndarray`):
-            The image to resize.
-        size (`int` or `Tuple[int, int]` or List[int] or Tuple[int]):
-            The size to use for resizing the image. If `size` is a sequence like (h, w), output size will be matched to
-            this.
-
-            If `size` is an int and `default_to_square` is `True`, then image will be resized to (size, size). If
-            `size` is an int and `default_to_square` is `False`, then smaller edge of the image will be matched to this
-            number. i.e, if height > width, then image will be rescaled to (size * height / width, size).
-        default_to_square (`bool`, *optional*, defaults to `True`):
-            How to convert `size` when it is a single int. If set to `True`, the `size` will be converted to a square
-            (`size`,`size`). If set to `False`, will replicate
-            [`torchvision.transforms.Resize`](https://pytorch.org/vision/stable/transforms.html#torchvision.transforms.Resize)
-            with support for resizing only the smallest edge and providing an optional `max_size`.
-        max_size (`int`, *optional*):
-            The maximum allowed for the longer edge of the resized image: if the longer edge of the image is greater
-            than `max_size` after being resized according to `size`, then the image is resized again so that the longer
-            edge is equal to `max_size`. As a result, `size` might be overruled, i.e the smaller edge may be shorter
-            than `size`. Only used if `default_to_square` is `False`.
-        input_data_format (`ChannelDimension`, *optional*):
-            The channel dimension format of the input image. If unset, will use the inferred format from the input.
-
-    Returns:
-        `tuple`: The target (height, width) dimension of the output image after resizing.
-    """
-    if isinstance(size, (tuple, list)):
-        if len(size) == 2:
-            return tuple(size)
-        elif len(size) == 1:
-            # Perform same logic as if size was an int
-            size = size[0]
-        else:
-            raise ValueError(
-                "size must have 1 or 2 elements if it is a list or tuple")
-
-    if default_to_square:
-        return (size, size)
-
-    height, width = get_image_size(input_image, input_data_format)
-    short, long = (width, height) if width <= height else (height, width)
-    requested_new_short = size
-
-    new_short, new_long = requested_new_short, int(
-        requested_new_short * long / short)
-
-    if max_size is not None:
-        if max_size <= requested_new_short:
-            raise ValueError(
-                f"max_size = {max_size} must be strictly greater than the requested "
-                f"size for the smaller edge size = {size}"
-            )
-        if new_long > max_size:
-            new_short, new_long = int(
-                max_size * new_short / new_long), max_size
-
-    return (new_long, new_short) if width <= height else (new_short, new_long)
-
-
-def get_image_size(image: np.ndarray, channel_dim: ChannelDimension = None) -> Tuple[int, int]:
-    """
-    Returns the (height, width) dimensions of the image.
-
-    Args:
-        image (`np.ndarray`):
-            The image to get the dimensions of.
-        channel_dim (`ChannelDimension`, *optional*):
-            Which dimension the channel dimension is in. If `None`, will infer the channel dimension from the image.
-
-    Returns:
-        A tuple of the image's height and width.
-    """
-    if channel_dim is None:
-        channel_dim = infer_channel_dimension_format(image)
-
-    if channel_dim == ChannelDimension.FIRST:
-        return image.shape[-2], image.shape[-1]
-    elif channel_dim == ChannelDimension.LAST:
-        return image.shape[-3], image.shape[-2]
-    else:
-        raise ValueError(f"Unsupported data format: {channel_dim}")
-
-
-def to_pil_image(
-    image: np.ndarray,
-    do_rescale: Optional[bool] = None,
-    input_data_format: Optional[Union[str, ChannelDimension]] = None,
-) -> "PIL.Image.Image":
-    """
-    Converts `image` to a PIL Image. Optionally rescales it and puts the channel dimension back as the last axis if
-    needed.
-
-    Args:
-        image (`PIL.Image.Image` or `numpy.ndarray` or `torch.Tensor` or `tf.Tensor`):
-            The image to convert to the `PIL.Image` format.
-        do_rescale (`bool`, *optional*):
-            Whether or not to apply the scaling factor (to make pixel values integers between 0 and 255). Will default
-            to `True` if the image type is a floating type and casting to `int` would result in a loss of precision,
-            and `False` otherwise.
-        input_data_format (`ChannelDimension`, *optional*):
-            The channel dimension format of the input image. If unset, will use the inferred format from the input.
-
-    Returns:
-        `PIL.Image.Image`: The converted image.
-    """
-
-    if isinstance(image, PIL.Image.Image):
-        return image
-
-    # If the channel has been moved to first dim, we put it back at the end.
-    image = to_channel_dimension_format(
-        image, ChannelDimension.LAST, input_data_format)
-
-    # If there is a single channel, we squeeze it, as otherwise PIL can't handle it.
-    image = np.squeeze(image, axis=-1) if image.shape[-1] == 1 else image
-
-    # PIL.Image can only store uint8 values so we rescale the image to be between 0 and 255 if needed.
-    do_rescale = _rescale_for_pil_conversion(
-        image) if do_rescale is None else do_rescale
-
-    if do_rescale:
-        image = rescale(image, 255)
-
-    image = image.astype(np.uint8)
-    return PIL.Image.fromarray(image)
-
-
-def to_channel_dimension_format(
-    image: np.ndarray,
-    channel_dim: Union[ChannelDimension, str],
-    input_channel_dim: Optional[Union[ChannelDimension, str]] = None,
-) -> np.ndarray:
-    """
-    Converts `image` to the channel dimension format specified by `channel_dim`.
-
-    Args:
-        image (`numpy.ndarray`):
-            The image to have its channel dimension set.
-        channel_dim (`ChannelDimension`):
-            The channel dimension format to use.
-        input_channel_dim (`ChannelDimension`, *optional*):
-            The channel dimension format of the input image. If not provided, it will be inferred from the input image.
-
-    Returns:
-        `np.ndarray`: The image with the channel dimension set to `channel_dim`.
-    """
-    if not isinstance(image, np.ndarray):
-        raise ValueError(
-            f"Input image must be of type np.ndarray, got {type(image)}")
-
-    if input_channel_dim is None:
-        input_channel_dim = infer_channel_dimension_format(image)
-
-    target_channel_dim = ChannelDimension(channel_dim)
-    if input_channel_dim == target_channel_dim:
-        return image
-
-    if target_channel_dim == ChannelDimension.FIRST:
-        image = image.transpose((2, 0, 1))
-    elif target_channel_dim == ChannelDimension.LAST:
-        image = image.transpose((1, 2, 0))
-    else:
-        raise ValueError(
-            "Unsupported channel dimension format: {}".format(channel_dim))
-
-    return image
-
-
-def _rescale_for_pil_conversion(image):
-    """
-    Detects whether or not the image needs to be rescaled before being converted to a PIL image.
-
-    The assumption is that if the image is of type `np.float` and all values are between 0 and 1, it needs to be
-    rescaled.
-    """
-    if image.dtype == np.uint8:
-        do_rescale = False
-    elif np.allclose(image, image.astype(int)):
-        if np.all(0 <= image) and np.all(image <= 255):
-            do_rescale = False
-        else:
-            raise ValueError(
-                "The image to be converted to a PIL image contains values outside the range [0, 255], "
-                f"got [{image.min()}, {image.max()}] which cannot be converted to uint8."
-            )
-    elif np.all(0 <= image) and np.all(image <= 1):
-        do_rescale = True
-    else:
-        raise ValueError(
-            "The image to be converted to a PIL image contains values outside the range [0, 1], "
-            f"got [{image.min()}, {image.max()}] which cannot be converted to uint8."
-        )
-    return do_rescale
-
-
-def rescale(
-    image: np.ndarray,
-    scale: float,
-    data_format: Optional[ChannelDimension] = None,
-    dtype: np.dtype = np.float32,
-    input_data_format: Optional[Union[str, ChannelDimension]] = None,
-) -> np.ndarray:
-    """
-    Rescales `image` by `scale`.
-
-    Args:
-        image (`np.ndarray`):
-            The image to rescale.
-        scale (`float`):
-            The scale to use for rescaling the image.
-        data_format (`ChannelDimension`, *optional*):
-            The channel dimension format of the image. If not provided, it will be the same as the input image.
-        dtype (`np.dtype`, *optional*, defaults to `np.float32`):
-            The dtype of the output image. Defaults to `np.float32`. Used for backwards compatibility with feature
-            extractors.
-        input_data_format (`ChannelDimension`, *optional*):
-            The channel dimension format of the input image. If not provided, it will be inferred from the input image.
-
-    Returns:
-        `np.ndarray`: The rescaled image.
-    """
-    if not isinstance(image, np.ndarray):
-        raise ValueError(
-            f"Input image must be of type np.ndarray, got {type(image)}")
-
-    rescaled_image = image * scale
-    if data_format is not None:
-        rescaled_image = to_channel_dimension_format(
-            rescaled_image, data_format, input_data_format)
-
-    rescaled_image = rescaled_image.astype(dtype)
-
-    return rescaled_image
-
-
-def get_image_size(image: np.ndarray, channel_dim: ChannelDimension = None) -> Tuple[int, int]:
-    """
-    Returns the (height, width) dimensions of the image.
-
-    Args:
-        image (`np.ndarray`):
-            The image to get the dimensions of.
-        channel_dim (`ChannelDimension`, *optional*):
-            Which dimension the channel dimension is in. If `None`, will infer the channel dimension from the image.
-
-    Returns:
-        A tuple of the image's height and width.
-    """
-    if channel_dim is None:
-        channel_dim = infer_channel_dimension_format(image)
-
-    if channel_dim == ChannelDimension.FIRST:
-        return image.shape[-2], image.shape[-1]
-    elif channel_dim == ChannelDimension.LAST:
-        return image.shape[-3], image.shape[-2]
-    else:
-        raise ValueError(f"Unsupported data format: {channel_dim}")
-
-
-def get_channel_dimension_axis(
-    image: np.ndarray, input_data_format: Optional[Union[ChannelDimension, str]] = None
-) -> int:
-    """
-    Returns the channel dimension axis of the image.
-
-    Args:
-        image (`np.ndarray`):
-            The image to get the channel dimension axis of.
-        input_data_format (`ChannelDimension` or `str`, *optional*):
-            The channel dimension format of the image. If `None`, will infer the channel dimension from the image.
-
-    Returns:
-        The channel dimension axis of the image.
-    """
-    if input_data_format is None:
-        input_data_format = infer_channel_dimension_format(image)
-    if input_data_format == ChannelDimension.FIRST:
-        return image.ndim - 3
-    elif input_data_format == ChannelDimension.LAST:
-        return image.ndim - 1
-    raise ValueError(f"Unsupported data format: {input_data_format}")
