+++ clip/preprocessing/image_constants.py
@@ -1,0 +1,5 @@
+"""
+Taken from https://github.com/huggingface/transformers/blob/main/src/transformers/utils/constants.py.
+"""
+OPENAI_CLIP_MEAN = [0.48145466, 0.4578275, 0.40821073]
+OPENAI_CLIP_STD = [0.26862954, 0.26130258, 0.27577711]
